+++ demisto_sdk/commands/common/content/tests/objects/pack_objects/indicator_type/indicator_type_test.py
@@ -4,7 +4,7 @@ from demisto_sdk.commands.common.constants import (INDICATOR_TYPES_DIR,
 from demisto_sdk.commands.common.content.objects.pack_objects import (
     IndicatorType, OldIndicatorType)
 from demisto_sdk.commands.common.content.objects_factory import \
-    ContentObjectFactory
+    path_to_pack_object
 from demisto_sdk.commands.common.tools import src_root
 
 TEST_DATA = src_root() / 'tests' / 'test_files'
@@ -14,7 +14,7 @@ INDICATOR_TYPE = TEST_CONTENT_REPO / PACKS_DIR / 'Sample01' / INDICATOR_TYPES_DI
 
 class TestIndicatorType:
     def test_objects_factory(self):
-        obj = ContentObjectFactory.from_path(INDICATOR_TYPE)
+        obj = path_to_pack_object(INDICATOR_TYPE)
         assert isinstance(obj, IndicatorType)
 
     def test_prefix(self):
@@ -25,7 +25,7 @@ class TestIndicatorType:
 class TestOldIndicatorType:
     @pytest.mark.parametrize(argnames="file", argvalues=["reputations.json"])
     def test_objects_factory(self, datadir, file: str):
-        obj = ContentObjectFactory.from_path(datadir[file])
+        obj = path_to_pack_object(datadir[file])
         assert isinstance(obj, OldIndicatorType)
 
     @pytest.mark.parametrize(argnames="file", argvalues=["reputations.json"])