+++ demisto_sdk/commands/common/content/tests/objects/pack_objects/layout/layout_test.py
@@ -2,7 +2,7 @@ import pytest
 from demisto_sdk.commands.common.content.objects.pack_objects import \
     LayoutsContainer
 from demisto_sdk.commands.common.content.objects_factory import \
-    ContentObjectFactory
+    path_to_pack_object
 from demisto_sdk.commands.common.tools import src_root
 
 TEST_DATA = src_root() / 'tests' / 'test_files'
@@ -12,7 +12,7 @@ TEST_CONTENT_REPO = TEST_DATA / 'content_slim'
 class TestLayoutsContainer:
     @pytest.mark.parametrize(argnames="file", argvalues=["layoutscontainer-Zimperium_event.json"])
     def test_objects_factory(self, datadir, file: str):
-        obj = ContentObjectFactory.from_path(datadir[file])
+        obj = path_to_pack_object(datadir[file])
         assert isinstance(obj, LayoutsContainer)
 
     @pytest.mark.parametrize(argnames="file", argvalues=["layoutscontainer-Zimperium_event.json"])