+++ CHANGELOG.md
@@ -8,9 +8,9 @@
 
 # v1.2.8
 * Added a new validation that checks in playbooks for the usage of `DeleteContext` in **validate** command.
+* Fixed an issue in the **upload** command where it would try to upload content entities with unsupported versions.
 * Added a new validation that checks in playbooks for the usage of specific instance in **validate** command.
 * Added the **--staged** flag to **validate** command to run on staged files only.
-
 
 
 # 1.2.7